fix(duplicate-events): addressed pr feedback
